requirements.txt
pandas
requests
plotly-express
plotly
python-dotenv
requests
sendgrid==6.6.0

Flask
app
